refactor(reportWebVitals): move it in global file for helpers
