+++ src/index.jsx
@@ -2,8 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
+
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './Helpers/reportWebVitals';
 import store from './store/reducer';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
